+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Chapter and Episode contain scode error
+- Counter: Scmd get typo
 
 ## [0.5.9-10] - 2020-06-23
 ### Added